fix(kaas): set kaas agent version at build time EE-2550 (#1522)
For the new KaaS feature. We need to deploy a specific version of the
agent to the cluster, normally this will be the matching release, so
portainer 2.13 will deploy agent 2.13 and so on.

However, during release testing the new version of the agent is not
live, yet the version for portainer has been increased. This causes
failures trying to deploy the agent. The environment variable
KAAS_AGENT_VERSION was used to override this, but our automated QA
builds seem unable to set environment variables.

Instead this allows overriding the version at build time with the 
`KAAS_AGENT_VERSION` variable. The variable and used in build_binary.sh,
build_binary_azuredevops.sh, and build_in_container.sh

IF YOU ARE BUILDING LOCALLY USING A DIFFERENT METHOD run a command
similar to this:

go build --ldflags "-X github.com/portainer/portainer-ee/api/kubernetes/cli.DefaultAgentVersion=latest" -o "../dist/portainer" ./cmd/portainer/