fix(license): allow replacing an existing license (#1681)
This allows users to override an existing license key (to replace a
corrupted key in the DB) by removing a check for duplicated licenses.

The check doesn't really make any sense because all that would happen if
we allow them to try is a request being made to the license server and
the same data being stored in the database. However, if the user's
database was somehow corrupted (perhaps during an old broken migration
script) then doing this will fix their database without manual
intervention.